+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Support pushing webhook events to message queues - #5940 by @patrys, @korycins
 - Adyen drop-in integration - #5914 by @korycins, @IKarbowiak
 - Add `change_currency` command - #6016 by @maarcingebala
+- Send a confirmation email when the order is canceled or refunded - #6017
 
 ### Breaking Changes